Info target to print information about makefiles
The phony target info prints target descriptions, a list of common
targets and more detailed information for the topic makefile that is
included first.  INFO is the list of targets for the first makefile
and INFOend may contain targets of all files that are shown at last.
All targets in the INFO list are ignored in the INFOend list.