Common macros for min, max and column/row numbers
The macros getmin, getmax, fcols and the programs NCOLS and NROWS
to count columns and rows in files are provided.  The macro fcols
uses NCOLS to count columns and subtracts IF_FUTURE then.

The variable IF_FUTURE shall be 0 if all columns are data and 1
if the last column denotes the ends of continuous trajectories.