information: Calculation of negentropy fixed
Subtract estimated entropy from the exact entropy of a Gaussian with the
same variance to obtain the estimated negentropy.  Before, a Gaussian
with the same variance centered with respect to the histogram ranges was
cut off at the range boarders and binned in the same histogram edges.
This procedure generates a density much more similar to the original
histogram than to a Gaussian function and underestimates the negentropy.